+++ CHANGELOG.md
@@ -6,8 +6,9 @@
   your error dialog now has problems, simply make `useLocalContext: true` to return to the old
   behavior.
 
-* Breaking change: The `StoreConnector.onInitialBuild` parameter now also gets the store. Where you
-  previously had `onInitialBuild(vm) {...}` now you have `onInitialBuild(store, vm) {...}`.
+* Breaking change: `StoreConnector.onInitialBuild` and `StoreConnector.onWillChange` parameters now
+  also get the context and the store. For example, where you previously
+  had `onInitialBuild(vm) {...}` now you have `onInitialBuild(context, store, vm) {...}`.
 
 # [9.0.9] - 2021/05/10